# specs/admin/openapi.yaml
openapi: 3.1.0
info:
  title: BWB Fiscal Admin API
  version: 0.1.10-draft
  description: >
    Contrato administrativo separado da API POS pública. Prefixo /admin/v1.
    Autenticação OIDC/JWT com RBAC (owner|admin|operator|auditor), distinta do
    Bearer credential_store POS. Runtime: fail_closed | injected (dev) |
    oidc_jwt (JWKS https, iss/aud exactos, role map; DEC-BO-003 / RM-BO-006).
    Segredos NÃO são devolvidos — zona SecAdm owner-only write-only (DEC-BO-001/002).
    0.1.1: PATCH scope-bindings (séries/timezone/ambiente metadados; RM-BO-002).
    0.1.2: ops visibility + audit list + secret-ref metadata sanitizado (RM-BO-003).
    0.1.3: matriz RBAC documentada; MFA adiado até IdP; sem ACL secret.reveal (RM-BO-004).
    0.1.4: SecAdm HTTP Put/Rotate/Revoke owner-only; respostas só metadados (RM-SECADM-003).
    0.1.5: listagens cadastro + PATCH status taxpayer/establishment (RM-BO-005).
    0.1.6: fundação OIDC/JWT provider-neutral documentada (RM-BO-006).
    0.1.7: observabilidade admin — health/ready/metrics + request_id (RM-BO-007).
    0.1.8: AuthorityProfile metadados HML/PRD (RM-AGTPREP-002 / DEC-BO-004); sem segredos; external_verified=false.
    0.1.9: UI SSR owner-only /admin/ui/authority-profiles (RM-AGTPREP-003); contrato HTTP inalterado; sem plaintext.
    0.1.10: SecAdm material multipart PEM/PKCS#12/credencial (RM-AGTPREP-004); password efémera; limites; ≠ AGT.
    Não pronto para produção. ≠ credenciais AGT. ≠ fornecedor IdP fixo.
  license:
...
servers:
...
security:
  - adminBearer: []
paths:
...
  /taxpayers:
    get:
      operationId: listTaxpayers
      summary: Lista contribuintes (metadados não secretos)
      tags: [Cadastros]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxpayerList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createTaxpayer
      summary: Cria contribuinte (cadastro não secreto)
      tags: [Cadastros]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaxpayerRequest"
      responses:
        "201":
          description: Contribuinte criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxpayer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /taxpayers/{taxpayer_id}:
    get:
      operationId: getTaxpayer
      summary: Obtém contribuinte por id
      tags: [Cadastros]
      parameters:
        - $ref: "#/components/parameters/TaxpayerId"
      responses:
        "200":
          description: Contribuinte
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxpayer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTaxpayerStatus
      summary: Actualiza status active/inactive
      tags: [Cadastros]
      parameters:
        - $ref: "#/components/parameters/TaxpayerId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Contribuinte actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Taxpayer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /establishments:
    get:
      operationId: listEstablishments
      summary: Lista estabelecimentos
      tags: [Cadastros]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: taxpayer_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstablishmentList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createEstablishment
      summary: Cria estabelecimento sob um contribuinte
      tags: [Cadastros]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEstablishmentRequest"
      responses:
        "201":
          description: Estabelecimento criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Establishment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /establishments/{establishment_id}:
    get:
      operationId: getEstablishment
      summary: Obtém estabelecimento por id
      tags: [Cadastros]
      parameters:
        - $ref: "#/components/parameters/EstablishmentId"
      responses:
        "200":
          description: Estabelecimento
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Establishment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateEstablishmentStatus
      summary: Actualiza status active/inactive
      tags: [Cadastros]
      parameters:
        - $ref: "#/components/parameters/EstablishmentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Estabelecimento actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Establishment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /scope-bindings:
    get:
      operationId: listScopeBindings
      summary: Lista scope bindings (metadados não secretos)
      tags: [Cadastros]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScopeBindingList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createScopeBinding
      summary: Cria binding de scope (metadados não secretos)
      tags: [Cadastros]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScopeBindingRequest"
      responses:
        "201":
          description: Binding criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScopeBinding"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /scope-bindings/{scope_id}:
    get:
      operationId: getScopeBinding
      summary: Obtém scope binding por scope_id
      tags: [Cadastros]
      parameters:
        - $ref: "#/components/parameters/ScopeId"
      responses:
        "200":
          description: Binding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScopeBinding"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateScopeBindingConfig
      summary: Actualiza séries/timezone/ambiente/status (metadados não secretos)
      description: >
        Não altera taxpayer_id nem establishment_id. Ambiente é metadado
        (homologation|production|development); não lê nem escreve segredos.
      tags: [Cadastros]
      parameters:
        - $ref: "#/components/parameters/ScopeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScopeConfigRequest"
      responses:
        "200":
          description: Binding actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScopeBinding"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /authority-profiles:
    get:
      operationId: listAuthorityProfiles
      summary: Lista perfis de autoridade (metadados; sem segredos)
      description: >
        DEC-BO-004 / RM-AGTPREP-002. ops.read. Nunca devolve PEM, passwords ou tokens.
        external_verified é sempre false até probe AGT real.
      tags: [Autoridade]
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuthorityProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createAuthorityProfile
      summary: Cria perfil de autoridade (owner-only; sem segredos)
      tags: [Autoridade]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuthorityProfileRequest"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorityProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /authority-profiles/{profile_id}:
    get:
      operationId: getAuthorityProfile
      summary: Obtém perfil de autoridade
      tags: [Autoridade]
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Perfil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorityProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: patchAuthorityProfile
      summary: Actualiza metadados/readiness do perfil (owner-only)
      description: external_verified não pode ser true por este endpoint.
      tags: [Autoridade]
      parameters:
        - name: profile_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAuthorityProfileRequest"
      responses:
        "200":
          description: Actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorityProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
  /audit-events:
    get:
      operationId: listAuditEvents
      summary: Lista eventos de auditoria admin (append-only, sem segredos)
      tags: [Ops]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Lista de eventos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditEventList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /ops/submissions:
    get:
      operationId: listOpsSubmissions
      summary: Visibilidade ops de submissões (estados; sem corpos secretos)
      tags: [Ops]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Resumos sanitizados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionSummaryList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OIDC/JWT de operador (RBAC: owner|admin|operator|auditor). Distinto do
        bearerAuth POS. Só owner acede SecAdm (e subject allowlist no runtime).
        Modos: fail_closed; injected (só development); oidc_jwt com JWKS https
        e mapa de grupos configurável (DEC-BO-003). Sessão browser ≠ este Bearer
        (RM-UI-005). Nunca enviar refresh tokens neste header.
...